crud with package and branch

+++ src/services/base.service.ts
@@ -1,5 +1,6 @@
-import { Model } from 'mongoose';
+import mongoose, { Model } from 'mongoose';
 import { ICRUDService } from '../utils/ICRUDService';
+import { BadRequestError } from '../errors/badRequestError';
 
 export abstract class BaseService<T> implements ICRUDService<T> {
   public readonly model: Model<T>;
@@ -23,6 +24,9 @@ export abstract class BaseService<T> implements ICRUDService<T> {
   }
 
   async getById(id: string): Promise<T | null> {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new BadRequestError('Invalid ID format');
+    }
     return await this.model.findById(id);
   }
 
@@ -31,7 +35,10 @@ export abstract class BaseService<T> implements ICRUDService<T> {
   }
 
   async update(id: string, data: Partial<T>): Promise<T | null> {
-    await this.beforeUpdate(id, data);
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new BadRequestError('Invalid ID format');
+    }
+    this.beforeUpdate(id, data);
     const document = await this.model.findById(id);
 
     if (!document) {
@@ -44,6 +51,9 @@ export abstract class BaseService<T> implements ICRUDService<T> {
   }
 
   async delete(id: string): Promise<void> {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new BadRequestError('Invalid ID format');
+    }
     await this.model.findByIdAndDelete(id);
   }
 }
